Move inner types into types directory
