+++ src/types/card.g.ts
